backend/requirements-dev.txt
# Web framework
fastapi==0.111.0
uvicorn[standard]==0.30.1

# Database + ORM
SQLAlchemy==2.0.31
alembic==1.13.2
psycopg2-binary==2.9.9   # PostgreSQL driver

# Config Management
pydantic==2.8.2
python-dotenv==1.0.1

# Testing
pytest==8.2.2
httpx==0.27.0             # Async client for FastAPI testing
coverage==7.6.1   # optional, for test coverage

# Optional dev tools
black==24.4.2
flake8==7.1.0
